Add connection to mongo, define a simple model and a router
When node start in backend/, a connection to mongo is made and we add
a router to the /product endpoint defined in /routes. This router
is using the models/product file and for now return an empty array
since we don't have yet any products in the fridge. A product has a name
and a quantity for the moment.